// aries-indeed_framework/backend/src/generic.modules/database/db.enum.ts
export enum databases {
    LoggedDataDB = 'LoggedDataDB',
    RheinenergieDataDB = 'RheinenergieDataDB',
    SMADataDB = 'SMADataDB',
    SMAprobandenDataDB = 'SMAprobandenDataDB',
    LiqwotecDataDB = 'LiqwotecDataDB',
    LEWDataDB = 'LEWDataDB',
    SchweigerDataDB = 'SchweigerDataDB',
    ffeDataDB = 'ffeDataDB',
    AggregatedDataDB = 'AggregatedDataDB',
    NotarizedDataDB = 'NotarizedDataDB',
    NotarizedOwnerDataDB = 'NotarizedOwnerDataDB',
    RemoteNotarizationDataDB = 'RemoteNotarizationDataDB',
    CollectiveNotarizationDataDB = 'CollectiveNotarizationDataDB',
    LoggedProsumerDataDB = 'LoggedProsumerDataDB',
    CommittedProsumerDataDB = 'CommittedProsumerDataDB',
    VerifiableProsumerDataDB = 'VerifiableProsumerDataDB',
    labeledConsumerAggregationDB = 'labeledConsumerAggregationDB',

    ConsumerConsumptionDB = 'ConsumerConsumption',
    ConsumerAggregationDB = 'ConsumerAggregation',
    ConsumerCompilationDB = 'ConsumerCompilation',
    ConsumerCompilationAggregationDB = 'ConsumerCompilationAggregation',
    ConsumerFootprintDB = 'ConsumerFootprint',
    ConsumerFootprintDataDB = 'ConsumerFootprintData',
    ConsumerForecastDB = 'ConsumerForecast',
    ConsumerPriorizationDB = 'ConsumerPriorization',
    masterRegistryDB = 'masterRegistry',
    usersDB='usersDB'
}